Extract recovery lifecycle into focused recovery_lifecycle module
